Tidy tests and remove selenium jar
