--- selenium-server-standalone-2
@@ -1,1 +1,0 @@
-<?xml version='1.0' encoding='UTF-8'?><Error><Code>NoSuchKey</Code><Message>The specified key does not exist.</Message></Error>
