+++ wdio.conf.js
@@ -10,7 +10,7 @@ exports.config = {
     // directory is where your package.json resides, so `wdio` will be called from there.
     //
     specs: [
-        './test/**/*.js'
+        './test/feature/**/*.js'
     ],
     // Patterns to exclude.
     exclude: [
@@ -68,7 +68,7 @@ exports.config = {
     //
     // Set a base URL in order to shorten url command calls. If your url parameter starts
     // with "/", then the base url gets prepended.
-    baseUrl: 'http://localhost:3000',
+    baseUrl: 'http://localhost:' + parseInt(process.env.NODE_PORT),
     //
     // Default timeout for all waitFor* commands.
     waitforTimeout: 10000,
